Code was moved from separated files to relevated one
