+++ OIIInspector/utils.py
@@ -1,5 +1,10 @@
 import argparse
-
+import subprocess
+import logging
+import shlex
+import logging
+import json
+import re
 
 def setup_arg_parser(args):
     """
@@ -36,3 +41,59 @@ def setup_arg_parser(args):
         holder.add_argument(*aliases, **kwargs)
 
     return parser
+
+
+def run_cmd(cmd, err_msg=None, tolerate_err=False):
+    """
+    Run a command locally.
+    Args:
+        cmd (str):
+            Shell command to be executed.
+        err_msg (str):
+            Error message written when the command fails.
+        tolerate_err (bool):
+            Whether to tolerate a failed command.
+
+    Returns (str, str):
+        Tuple of stdout and stderr generated by the command.
+    """
+    log = logging.getLogger("OIIInspector")
+    err_msg = err_msg or "An error has occurred when executing a command."
+    p = subprocess.Popen(shlex.split(cmd), stdout=subprocess.PIPE)
+    out, err = p.communicate()
+    if p.returncode != 0 and not tolerate_err:
+        log.error("Command {0} failed with {1}".format(cmd, err))
+        raise RuntimeError(err_msg)
+
+    return out.decode('utf-8')
+
+
+def convert_output(api_response):
+    """
+    Converts answer from API into JSON
+    Args:
+        api_response (str):
+            String received from grpcurl call on api.Registry
+
+    Returns (str)
+        String already convertible into JSON format.
+    """
+
+    log = logging.getLogger("OIIInspector")
+    if api_response == "":
+        log.error("Answer from API is empty")
+        raise RuntimeError("Answer from API is empty")
+
+    api_response = re.sub(r'}\s*{', "} ,{", api_response)
+    json_data = json.loads(api_response)
+    if 'csvJson' in json_data:
+        json_data['csvJson'] = json.loads(json_data['csvJson'])
+
+    if 'object' in json_data:
+        new_object = []
+        for item in json_data['object']:
+            new_object.append(json.loads(item))
+        json_data['object'] = new_object
+    if 'spec' in json_data:
+        json_data['spec'] = json.loads(json_data['spec'])
+    return json_data
